Keep analytics QA stable with MPP enabled
Run analytics REST QA with MPP enabled by default and add a strict
property for exercising both kill-switch states.

Preserve no-join CBO plans until unary aggregate worker tiers are
supported, and stabilize expectations exposed by the main sync.

+++ sandbox/qa/analytics-engine-rest/src/test/java/org/opensearch/analytics/qa/SystemFunctionsIT.java
@@ -76,9 +76,10 @@ public class SystemFunctionsIT extends AnalyticsRestTestCase {
         assertFirstRowString(oneRow("key00") + "| eval t = typeof(datetime0) | fields t", "TIMESTAMP");
     }
 
-    /** {@code typeof} on an arithmetic expression. {@code int0 * 2} stays INT. */
+    /** {@code typeof} on an arithmetic expression. PPL integer literals are BIGINT, so
+     *  {@code int0 * 2} is widened to BIGINT by Calcite's arithmetic type inference. */
     public void testTypeofArithmetic() throws IOException {
-        assertFirstRowString(oneRow("key00") + "| eval t = typeof(int0 * 2) | fields t", "INT");
+        assertFirstRowString(oneRow("key00") + "| eval t = typeof(int0 * 2) | fields t", "BIGINT");
     }
 
     // ── helpers ─────────────────────────────────────────────────────────────
